feat: RSS-PZ-10 Enable word cards reversible interaction
