+++ src/components/puzzleContainer.ts
@@ -1,4 +1,4 @@
-import ElementCreator from '../../utils/elementСreator';
+import ElementCreator from '../utils/elementСreator';
 
 export const puzzleContainer = new ElementCreator({
   tag: 'div',
